lang_chain_basics/requirement.txt
langchain-openai
python-dotenv
chromadb
langchain-chroma
langchain_community
langchain